complete: user rank list

+++ src/main/java/cn/edu/sdtbu/model/vo/user/UserSimpleInfoVO.java
@@ -16,5 +16,5 @@ public class UserSimpleInfoVO extends BaseUserVO {
     String school;
     String email;
     UserRole role;
-    Integer rank;
+    Long rank;
 }
